+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - [TN] [ITN] [te] Add Telugu text normalization and inverse text normalization grammars
 - [TN] [ITN] [ml] Add Malayalam text normalization and inverse text normalization grammars
 - [TN] [ITN] [kn] Add Kannada text normalization and inverse text normalization grammars
+- [TN] [ITN] [hi] Add Hindi text normalization and inverse text normalization grammars